feat(ssa-workflow)!: use standalone canonical key crate
Move fixed-width canonical key encoding into ssa-canonical-key and replace the unsafe buffer contract with a checked CanonicalWriter and reusable typed CanonicalBuffer.

BREAKING CHANGE: CanonicalEncode implementations now define encode with CanonicalWriter, and execute_one_with_buffer accepts CanonicalBuffer instead of a raw byte slice.
