+++ crates/ssa-canonical-key/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "ssa-canonical-key"
+version = "0.1.0"
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+
+[lints]
+workspace = true
