+++ crates/ssa-workflow/Cargo.toml
@@ -17,6 +17,7 @@ parking_lot = { workspace = true }
 postcard = { workspace = true, optional = true }
 rand = { workspace = true }
 rayon = { workspace = true }
+ssa-canonical-key = { workspace = true }
 serde = { workspace = true, optional = true }
 thiserror = { workspace = true }
 zstd = { workspace = true, optional = true }
