+++ crates/ssa-workflow/src/cache/mod.rs
@@ -9,12 +9,12 @@ macro_rules! warn {
 macro_rules! warn {
     ($($arg:tt)*) => { eprintln!($($arg)*) };
 }
-pub use canonical_encode::{CanonicalEncode, CanonicalEncodeWriter};
 pub use encoded::EncodedCache;
 #[cfg(feature = "lru")]
 pub use memory::ManagedLruCache;
 pub use memory::{ManagedHashCache, ManagedMemoryCache};
 pub use provider::{CacheProvider, PersistentCacheProvider, StorageProvider, StorageProviderExt};
+pub use ssa_canonical_key::{CanonicalBuffer, CanonicalEncode, CanonicalWriter};
 
 use crate::Result;
 
@@ -69,7 +69,6 @@ impl CloneShared for () {
     fn clone_shared(&self) -> Self {}
 }
 
-mod canonical_encode;
 pub mod codec;
 mod encoded;
 pub mod memory;
