+++ crates/ssa-workflow/tests/cache_integration.rs
@@ -8,7 +8,7 @@ use std::sync::{
 use ssa_workflow::{
     Compute,
     cache::{
-        Cache, CacheProvider, CanonicalEncode, CloneShared, EncodedCache,
+        Cache, CacheProvider, CanonicalBuffer, CloneShared, EncodedCache,
         codec::{CheckedCodec, CloneFresh, CodecEngine, Error as CodecError, SkipReason},
         storage::EncodedStorage,
     },
@@ -165,8 +165,8 @@ fn encoded_cache_treats_checked_corruption_as_miss_in_execution_flow() -> Result
     assert_eq!(compute.execute_one(3usize)?, 30);
     assert_eq!(compute_calls.load(Ordering::SeqCst), 1);
 
-    let mut key_buffer = vec![0u8; usize::SIZE];
-    let key = unsafe { 3usize.encode_with_buffer(&mut key_buffer) }.to_vec();
+    let mut key_buffer = CanonicalBuffer::new();
+    let key = key_buffer.encode(&3usize).to_vec();
     let mut corrupted = store_handle
         .fetch_encoded(&key)?
         .expect("value should have been stored")
